iOS: Very basic Button implementation (text only). Implemented bare minimum for iup font to make the code work.
Had trouble with UIButton sizeToFit. Currently trying to use boundingRectWithSize and sizeThatFits to size the button. This is kind of working, but IUP later calls back into iupdrvFontGetMultiLineStringSize and we seem to be resizing for this same label. And for some reason, if I put in the previous button's bounds as input, the resulting size comes a lot smaller than before. I'm currently using screen size, but that is smaller too (not as small). I'm confused by this.

System font and font size are hard coded right now. This needs to be fixed.

Removed old Mac code from font file.