iOS: Implemented proposal for new IUP startup sequence because the existing one cannot work on Android and iOS.
On iOS, there is a problem because IUP APIs are expected to work before calling IupMainLoop(). But in order to implement a functionality, I need the iOS app to be fully formed (e.g. I need the app delegate to have been created, also autorelease pool, etc.) To properly start up the application, I must use UIApplicationMain, but that takes control of the event loop and doesn't return (hence why it was put in IupMainLoop).

Ultimately, the realization is that this is effectively the same problem Android has, just with different technicalities. So IUP's startup sequence must be changed. And I believe it is possible to make the legacy design conform to the new way, but it is impossible to make the legacy design work on Android and iOS.

The new convention looks for a C function called IupEntryPoint.
The code will look for this entry. If not found, it will fallback to IupEntryPoint(). The symbol is looked up by name using dlsym to avoid creating a linker dependency between the Iup library and the main library.

(The user may also specify a callback function by calling
IupSetFunction("ENTRY_POINT", (Icallback)IupEntryPoint);
This is how I expect other platforms to set/change the callback. But on Android, this code path is pretty impractical.)

So for a typical cross-platform project, I expect something like:

// TODO: Proposed exit callback which the user should setup in the IupEntryPoint() callback.
void IupExitCallback()
{
	IupClose();
}

// This is where user code should start
void IupEntryPoint()
{
	// TODO: Proposed exit callback so the user can call IupClose()
	IupSetFunction("EXIT_CB", (Icallback)IupExitCallback);

	Ihandle* button = IupButton("Iup Button", "");
	IupSetCallback(button, "ACTION", (Icallback)OnButtonCallback);
	Ihandle* dialog = IupDialog(button);
	IupShow(dialog);
}

// Remember that main is never run on Android.
// This should be a template/copy-and-paste for all IUP projects.
// Don't put any user code here. All user code should start in IupEntryPoint.
int main(int argc, char* argv[])
{
	IupOpen(0, NULL);
	IupSetFunction("ENTRY_POINT", (Icallback)IupEntryPoint);
	// User can no longer assume that IupMainLoop blocks, and may return immediately.
        // You should probably not put code after this call
	IupMainLoop();
	return 0;
}