Cocoa: Improved NSWindow creation behavior to cascade correctly when new windows are created. The very first window will call center. Also fixed a bug where closing a window would quit the app even if multiple windows were open.
The bug was caused by the interaction of iupDialogShowXY and window visibility. In Cocoa when the window is being created, it returns true for isVisible. Iup seems to assume because it is not on screen yet, it will return false. Iup skips its initialization code, which includes a counter of how many windows are open. When I close a first window, Iup decrements its counter and thinks all windows are closed. It will then call IupExitLoop() to quit the program which is wrong. I have a hack in place to avoid this. We might need to discuss a better solution.