Cocoa: This is a hideous hack trying to get vertical progress bars.
setCenterFrameRotation seems to be broken. (A Google search suggests it broke in El Capitan.)
I tried modifying the CALayer's transform directly. This kind of worked, but flickered like crazy and reverted to horizonital when window resizing.
So I tried to set the CALayer's drawing delegate to set the transform then. But now it doesn't show.

I'm planning to backout most of these, but am committing in case I want to look at this again in the future.