Cocoa: Slightly simplified from the previous check in, this is an attempt to get vertical progress bars.
This sets the transform directly on the CALayer of the NSView.
This mostly works, but on window resize, the NSView reasserts itself and clobbers my transform. The previous check in tried to be clever and use the CALayerDelegate to reassert my transform, but I just broke the renderering.
Extra note: I didn't have to enable layer backed views for this to work. (Sierra)

So I think I'm going to back this out or disable it.

I think filing a bug to get Apple to fix setFrameCenterRotation is the best solution.