Cocoa: Initial IupList implementation. Dropdown (NSPopupButton) and Editbox+Dropdown (NSComboBox)
Fonts were hacked/fixed to make this work, but it still needs serious attention. The new implementaiton It is potentially LEAKING when it registers a new font, but it is kind of cache/lookup, so it doesn't really ever free.

Most of the list callbacks and attributes are not implemented.
Will look at NSTableView for MULTIPLE.
There is no obvious native widget for plain Editbox.