Cocoa: Still trying to get layout working. Realized the attempt to use actual widget sizing in iupcocoa_font won't work because it will sometimes go through the string sizing and other times the widget sizing depending on whether the widget has been instantiated yet. This creates a problem down stream because the padding amounts don't know which metric was used initially. Added NSFrame (NSBox) to the mix.
In a bad actual case, resizing the view was causing the padding on the button to ever increase because the first pass was using string sizing, and then every resize used the real button and the padded on top of that which made it grow out of control.

NSBox is causing a few challenges. First, we need to get access to the inner container. Iup has an optional callback for InnerContainer which I can use. Alternatively, I can directly try to look for NSBox. I tried both. For the former, I'm concerned that some of the cocoaCommon code might get confused because I was not expecting this when I wrote it. But on a little more scrutiny, I'm starting to think it works. If it does, then I like it better than the special NSBox cases I added (currently commented out).

Second, I need to tell IUP that the padding/inset for the NSBox. It looks like IUP did not anticipate this, so I am going to have to change the internals. IUP currently hardcodes the inset to 5.