IupCoreChanges: Refactoring IupFrame internals so the box size can be customized by each platform. Introduces iupdrvFrameGetDecorSize and iupdrvFrameGetTitleHeight. Also changed iupdrvFrameGetDecorOffset & iupdrvFrameHasClientOffset to pass Ihandle.
Mac/Cocoa required this change because NSBox padding is much bigger and has different values depending on the box label mode.
Always passing the Ihandle is useful because Cocoa needed to query whether the title text was set or not, but since IUP may not instantiate the widget first, there is no way to get know which mode the widget is without the Ihandle.

In general, I recommend all internal APIs always pass the Ihandle* ih, just to pre-empt platform specific cases like this.