Cocoa: Tricky main menu code that will insert new menu entries to the left of the "Window" entry.
Cocoa has strong conventions about what should be in the menu and where.
Currently I'm operating on the assumption that we are going to pre-populate a default menu for IUP and the user is going to add (and maybe remove) items.
So we need to search through the existing menu and determine where things go.
Current assumption: All normal menu categories are already in the menu. So if the user adds a new one, we must put it in the right place.
The Apple Human User Interface Guidelines (HIG) state that new menu entries appear between the View and Window items.
View is also sometimes optional, so for robustness, we should scan for Window and insert right before Window.
(This also handles the case where user entries have already been added since we will add after those entries which is expected behavior.)

TODO: Do something similar for the individual standardized submenu entries?