Cocoa: Replaced the default/magicial Open Recent menu in the XIB with a standard/generic submenu by the same name.
I don't think we can conform to Apple's expected behavior to make the Open Recents automatically work. To work best, we need NSDocumentController which we are not using.
Without it, it looks like it is a pretty nasty hack which requires private API.
And even without the private API, I don't know how it would fit with the design of IUP, unless IUP were to introduce specific API for this.

http://lapcatsoftware.com/blog/2007/07/10/working-without-a-nib-part-5-open-recent-menu/