Cocoa: Support for IupPopup (modal windows).
The runloop problem became an issue again.

This implementation uses ic->DlgPopup to declare that the system handles the modal/event-loop stuff which avoids IUP trying to call IupMainLoop again and breaking stuff. My original implementation hacked the core to not call IupMainLoop and IupExitLoop, but this seems better.

But this still exposed a bunch of problems with not using the official Cocoa [NSApp run] runloop. Creating a popup before the runloop starts sometimes causes a ghost/stuck window to remain after dismissed. I introduced a hack that will start the official runloop with a custom app delegate just for this one case and then stop it when done.