Cocoa: Did some memory clean ups. I seem to have a bug with the menu tear down. It crashes during IupClose.
IupClose()
In iupNamesDestroyHandles

static Ihandle* iNameGetTopParent(Ihandle* ih)
{
  Ihandle* parent = ih;
  while (parent->parent)  // <-- crashes here
    parent = parent->parent;
  return parent;
}

My current guess is that there is a connection between the initial window and the menu because on other platforms, they are connected. But on Mac, these are independent, but I am currently cheating and reusing the API that connects the two. So my first guess is that IUP knows it has extra Ihandles that have not been destroyed yet. But these menu handles have pointers to the window which I already destroyed. Hence a crash.