Cocoa: Currently fighting IupText sizing & layout issues. It appears GridBox and NORMALIZE can force growing the sizes which screws up the textbox.
The custom layout I wrote for Text masks the problem. While the layout forces the text to be the correct height, the widget gets mis-centered because the IUP position is not coordinated with how the layout constrained the height from increasing.

I have currently disabled the custom layout for text, which shows that if the TextField is allowed to grow, then the alignment is centered.

I believe NSButton has a related problem, but it is less noticable because Cocoa ignores attempts to increase the button height. This leads to an offset problem too, but I think it is less noticable since the Button is so much taller to begin with and the NORMALIZE is probably adjusting to accomodate the button since it is the largest widget.