Cocoa: Waypoint: Workaround for Mac bug causing label to render in the wrong place due to the custom NSTextFieldCell. A bunch of this code needs to be cleaned up and removed, but it contains the fundamental solution.
In creating our own subclass for NSTextFieldCell to provide vertical alignment,
we have hit a Mac bug.
When we first create a multi-line word-wrapped label that is top-aligned,
Mac seems to draw it in the wrong place.
Even though our overrides in our subclass are producing the correct values,
on the very initial creation, the cell is positioned too low (about midway).
Calling setNeedsDisplay:YES, updateCell:, and all sorts of things to trigger a redraw/relayout do nothing.
It almost feels like that Mac is caching the image and some other place in the renderer is putting it in the wrong position
because the parts I tried overriding in the cell subclass seem to be producing all the correct values.
The one thing that seems to fix the position is clicking on the label after it has been displayed.
And when I click on it, nothing in I can see in the Cell, e.g. drawInteriorWithFrame get triggered, which is another reason it feels like it is some deferred Mac rendering bug.
So, as a workaround, we need to simulate a mouse-click on the label when first displayed to force Mac to draw it in the correct place.
Simulating the mouse-click caused its own problems.
mouseDown: warns that the parameter needs to be non-nil. However, in my attempts to create a proper NSEvent to pass,
I discovered that I would only sometimes correct the bug, and other times, the call seems to enter an event loop and never returns and never creates a window.
Or sometimes nothing would happen at all.
I discovered calling performSelector:withObject:afterDelay with some delay increased chances of it working.
However, using a debugger/breakpoints would frequently break things.
Eventually, I tried just passing nil as the events, and everything worked, and without needing a delay.