Cocoa: Putting NSTextField in NSStackView to keep IupGrid from changing the height of the text field. Contains #ifdef __APPLE__ because I was told NSStackView is not available on GNUStep.
I tried making a similar solution using autoresizingMask, but could not get it to work. It is broken enough that the old approach of letting IupGrid change the size is better. I left the code, just in case somebody knows how to fix it.

The NSStackView uses a 10.11 API. In my side testing, I needed this for a contrived case I hit using IB. However, I'm not sure if we actually will hit that case. We might be able to use respondToSelector if we care about 10.9/10.10.