CMake: Changes to support building iupweb on Android (plus tweaks for Apple). Also passing -Wno-invalid-source-encoding
The module stuff is messy and kind of broken due to Android Studio’s support being kind of broken. I’m trying to build separate .aar’s for iup and iupweb. But Android Studio wants to duplicate everything when I build iupweb. I have a manual filter that avoids copying the libiup.so, etc, into the iupweb .aar, but some of the other resources are still being copied in despite my attempts to prevent it. The CMake integration is annoying too because in order to build iupweb, it needs to rebuild its own copy of iup again. And since it uses the same CMakeLists.txt, iup gets all the iupweb stuff. So there is a argument switch that controls building the iupweb part so it can be suppressed when building the core iup.

For -Wno-invalid-source-encoding, I had to use CMAKE_C_FLAGS instead of IUP_C_FLAGS because the latter was having no effect.