GTKDLOpen: Started implementing a runtime dlopen for libwebkitgtk for iupweb because it turns out to be even harder than usual to ship a single Linux binary that is expected to work on many distros. In addition to the typical library name differences, the indirect dependencies seem to be causing me more anguish than usual. And trying to bundle/distribute your own binaries is riskier than usual because web components have so many security problems. So shipping a lowest-common denominator libwebkit (and with all its dependencies), is problematic because it will be out of date, probably having a lot of security issues, not to mention missing a lot of modern features.
So the intent of this is to find the libwebkit (and its dependencies) on the user's system at runtime so you don't have to ship them. IupWebBrowserOpen will return IUP_ERROR if the library cannot be found. Additionally, it will call IupSetGlobal with key _IUP_WEBBROWSER_MISSING_DLL, with a string describing which library could not be found (libwebkitgtk-3.0.so or libwebkitgtk-1.0).so since it depends on whether you are using GTK2 or GTK2.