IupPostMessage: Prototype implementation of IupPostMessage() for Windows and GTK as proposed on the mailing list.
This feature introduces an API akin to Window's PostMessage API, but for cross-platform use. Its primary benefit is that it can solve the problem of being on a background thread, but needing to trigger a UI change on the main UI thread. Since all the native backends are not thread-safe, but many libraries use threads, this can be a show-stopping problem.

Because threads are ubiquitous, every platform provides ways to safely dispatch code back to the main UI thread. These are native mechanisms that are tied with the native event loops.

However, because IUP wraps/abstracts all the platforms' native event loops (e.g. IupMainLoop()), but it does not wrap any of these native mechanisms for dealing with threads, this is a serious limitation in IUP.

So this patch implements the easiest solution to solve this problem and provides a way to add an event to the native event queue and get it called back on the main UI thread.

This patch should not change any existing behavior, nor add any new dependencies. It does not even require threads or locks.

This patch can also be used for single-threaded scenarios, for those who simply wish to add future events to be processed on the next event loop. Async patterns tend to utilize this kind of mechanism.

Currently the API is unfinished and is expected to be changed.
To setup a callback handler:
  IupSetCallback(ih_button, "POSTMESSAGE_CB", (Icallback)OnCallback);
To post a message:
 void IupPostMessage   (Ihandle* ih, char*, void* message_data, int);

Currently the char* and int are unused, and only done because I didn't want to add anything to iupcbs.h until we decided on a final API.

I also changed "UITHREAD" to POSTMESSAGE_CB from my initial mailing list proposal.

A single-threaded test program called postmessage.c is included in this patch.