Cocoa: Tooltip implementation. Will only support TIP attribute and nothing else, not even the callback.
This uses NSView's native built-in tooltip API. It is a barebones simple API and Apple controls everything about the presentation and behavior (by design).
So none of the other IUP properties are supportable. There are no tunable properties and no callback notification for when a tip is displayed.

I've deleted the original implementor's TooltipWindow class. It is a bit of a shame to throw away this work, but part of the goal of IUP is to use native UI. Apple bakes in specific usability behaviors which can change between OS releases, so we don't want to be in a position of trying to emulate them if we don't have to. Also, there is a huge amount of code here compared to the one-liner we need to use the native API.
The original code has special cases for NSTextField, but I don't understand why. I also don't fully understand why the original implementor went to this trouble when the NSView tooltip API appears to be available since OS X 10.0.