Cocoa: New proposed attribute "CONTEXTMENU". Rather than forcing users to manually implement mouse click events and figure out which widget is underneath to produce a relavent context menu, Cocoa lets you setup a context menu for every NSResponder. So if the contextual menu is invoked (could be Ctrl-click, right mouse click if you have a button, trackpad two-finger touch if configured, etc), the menu you set for that specific view instance will display your menu. And the developer doesn't have to emulate all the different ways the user can invoke the menu.
Additionally, some widgets come with built-in defaults for the context menu (e.g. NSTextView which has copy/paste, dictionary, Google search, social media sharing, spell check, etc). So trying to manually create a generic popup menu that is not connected to the underlying widget loses this information. So we really want to encourage developers to use the native API.

CONTEXTMENU does just this. It allows you to create a menu (through the normal IupMenu APIs) and attach it as an attribute on the Ihandle of any widget. Additionally, if the widget provides a default set of options, a separator is added under the developer's items, and the built-in defaults are copied/appended to the list.

There is a common base routine in iupcocoa_common that is expected to be leveraged for most cases.

Currently implemented and tested: Label (text and image), Button
Currently implemented: Canvas, Toggle, ProgressBar
Not yet implemented: List, Text, Tree

Text looks hard/tricky because we must deal with the Field Editor.
NSTableView and NSOutlineView allows for every cell to have its own menu, so this needs to be thought about/designed.