Cocoa: Improvements to Canvas/Draw. Filled in most of the functions. I am able to get enough to see FlatButton (but not perfectly). Text is still broken, but now renders some text. (Font system is a major blocker.) Also found a retain count bug in how I was loading images. IUP caches images and since I was autoreleasing after ever load, cached values were getting a double autorelease since IUP does not know it should call retain or copy. For now, I no longer attempt to autorelease and am trusting Iup will release when it cleans it cache.
There is also some wonkiness with the coordination between Canvas and Draw. I created a way to pass a dc back to the View. I'm not happy with it, but all my other solutions were worse so far. Part of the problem is there is no good way to get the updated width/size values to the dc when it happens in the view. Also, I have become paranoid about the possibility that the CGContext gets destroyed and recreated so trying to hang on to the pointer may be bad. This design can help avoid that problem by always re-assigning the CGContext to the dc in the NSView at the key places.

Needed to move struct IdrawCanvas to separate header file so Canvas and Draw can share.

Also added iupimglib, iupim to Xcode project. The latter removed all files because it has a CD dependency.