Cocoa: CanvasView: Trying to get key focus to work. nextKeyView is broken; the recalculateKeyLoop is not working for unknown reasons. Also, I am having doubts IupControls should get focus rings (key focus) because of this paragraph in the Apple doc about needsPanelToBecomeKey:
Some views (for example, text fields) want to receive the keyboard focus when you click in them.
Other views (for example, buttons) receive focus only when you tab to them.
You wouldn't want focus to shift from a textfield that has editing in progress simply because you clicked on a check box.

Sooo... since IUP uses this mostly for buttons and not text entry, it seems like we should return NO.
But this means that the widgets will never get the focus ring.