Cocoa: TRAY: Re-did TRAY system so it is no longer a global singleton, but a per-instance ih. TRAY=YES is a create-only attribute.
I misunderstood the usage case. It now seems clear to me that TRAY is not shared with the Window instance. So as long as these remain separate, this will work.

I needed to add a bunch of type checks in the implementation since many code paths are shared, but these objects are completely unrelated & incompatible.

It is also worth noting that I need to explicitly NULL out HIDETASKBAR as the default implementation was making the status icon disappear. I don't understand what HIDETASKBAR is supposed to do, but I think it is non-sensical for Cocoa.